// src/interfaces/ChosenGames.d.ts
export interface ChosenGames {
  demonsSouls: {
    name: "Demon's Souls",
    isChosen: boolean
  },
  darkSouls1: {
    name: "Dark Souls 1",
    isChosen: boolean
  },
  bloodborne: {
    name: "Bloodborne",
    isChosen: boolean
  },
  darkSouls2: {
    name: "Dark Souls 2",
    isChosen: boolean
  },
  eldenRing: {
    name: "Elden Ring",
    isChosen: boolean
  },
  darkSouls3: {
    name: "Dark Souls 3",
    isChosen: boolean
  },
}